Authenticate and authorize the user to delete post

+++ controllers/api/v1/posts_api.js
@@ -1,53 +1,52 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
-
-module.exports.index = async function(req,res){
+module.exports.index = async function(req, res){
 
 
     let posts = await Post.find({})
         .sort('-createdAt')
         .populate('user')
         .populate({
-            path : 'comments',
-            populate : {
-                path : 'user'
+            path: 'comments',
+            populate: {
+                path: 'user'
             }
         });
 
-    return res.json(200,{
-        message : "List of posts",
-        posts : posts
-    });
+    return res.json(200, {
+        message: "List of posts",
+        posts: posts
+    })
 }
 
-module.exports.destroy = async (req,res)=>{
 
+module.exports.destroy = async function(req, res){
 
     try{
         let post = await Post.findById(req.params.id);
-        // if(post.user == req.user.id){
+
+        if (post.user == req.user.id){
             post.remove();
-            await Comment.deleteMany({post:req.params.id});
 
-            
+            await Comment.deleteMany({post: req.params.id});
 
-            return  res.json(200,{
-                message : "Post Deleted successfully."
+
+    
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully!"
             });
+        }else{
             
-        //}else{
-        //     req.flash('Error','You Cannot Delete This Post!');
-
-        //     res.redirect('back');
-        // } 
+            return res.json(401,{
+                message : "You cannot delete the Post!"
+            });
+        }
 
     }catch(err){
-        
-       console.log('*****',err);
-        return res.json(500,{
-            message : "Internal server error"
+        console.log('********', err);
+        return res.json(500, {
+            message: "Internal Server Error"
         });
-
     }
     
 }
